django/src/requirements.txt
asgiref==3.3.1
autopep8==1.5
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
django-allauth==0.44.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
easywatch==0.0.5
entrypoints==0.3
flake8==3.8.4
GDAL==3.0.4
hidpidaemon==18.4.6
httplib2==0.14.0
idna==2.8
Jinja2==2.11.2
kernelstub==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
pathtools==0.1.2
pep8==1.7.1
pop-transition==1.0.1
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pydbus==0.6.0
pyflakes==2.2.0
pyflatpak==0.4.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-debian===0.1.36ubuntu1
python-xlib==0.23
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
repoman==1.2.2
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sessioninstaller==0.0.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
staticjinja==0.4.0
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
wadllib==1.3.3
watchdog==0.10.4
xkit==0.0.0
